@@ -1,3 +1,3 @@
 module serve-socket
 
-go 1.16
+go 1.17
