@@ -1,5 +1,5 @@
-// +build acceptance
-// +build darwin,amd64
+//go:build acceptance && darwin && amd64
+// +build acceptance,darwin,amd64
 
 package os
 
