@@ -1,5 +1,5 @@
-// +build acceptance
-// +build darwin,arm64
+//go:build acceptance && darwin && arm64
+// +build acceptance,darwin,arm64
 
 package os
 
